Refs #5371. Move UDP and TCP transport under IP domain.
