+++ src/c/profile/transport/ip/tcp/tcp_transport_windows.c
@@ -1,31 +1,23 @@
-#include <uxr/client/profile/transport/tcp/tcp_transport_linux.h>
+#include <uxr/client/profile/transport/tcp/ip/tcp_transport_windows.h>
 #include "tcp_transport_internal.h"
 
-#include <arpa/inet.h>
-#include <unistd.h>
-#include <string.h>
-#include <errno.h>
-#include <signal.h>
-
-#ifdef PLATFORM_NAME_LINUX
-static void sigpipe_handler(int fd)
-{
-    (void)fd;
-}
-#endif
+#include <uxr/client/util/time.h>
 
 bool uxr_init_tcp_platform(struct uxrTCPPlatform* platform, const char* ip, uint16_t port)
 {
     bool rv = false;
 
-    /* Socket initialization. */
-    platform->poll_fd.fd = socket(PF_INET, SOCK_STREAM, 0);
-    if (-1 != platform->poll_fd.fd)
+    /* WSA initialization. */
+    WSADATA wsa_data;
+    if (0 != WSAStartup(MAKEWORD(2, 2), &wsa_data))
     {
-#ifdef PLATFORM_NAME_LINUX
-        signal(SIGPIPE, sigpipe_handler);
-#endif
+        return false;
+    }
 
+    /* Socket initialization. */
+    platform->poll_fd.fd = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    if (INVALID_SOCKET != platform->poll_fd.fd)
+    {
         /* Remote IP setup. */
         struct sockaddr_in temp_addr;
         temp_addr.sin_family = AF_INET;
@@ -40,14 +32,15 @@ bool uxr_init_tcp_platform(struct uxrTCPPlatform* platform, const char* ip, uint
         int connected = connect(platform->poll_fd.fd,
                                 &platform->remote_addr,
                                 sizeof(platform->remote_addr));
-        rv = (0 == connected);
+        rv = (SOCKET_ERROR != connected);
     }
     return rv;
 }
 
 bool uxr_close_tcp_platform(struct uxrTCPPlatform* platform)
 {
-    return (-1 == platform->poll_fd.fd) ? true : (0 == close(platform->poll_fd.fd));
+    bool rv = (INVALID_SOCKET == platform->poll_fd.fd) ? true : (0 == closesocket(platform->poll_fd.fd));
+    return (0 == WSACleanup()) && rv;
 }
 
 size_t uxr_write_tcp_data_platform(struct uxrTCPPlatform* platform,
@@ -56,8 +49,8 @@ size_t uxr_write_tcp_data_platform(struct uxrTCPPlatform* platform,
                                    uint8_t* errcode)
 {
     size_t rv = 0;
-    ssize_t bytes_sent = send(platform->poll_fd.fd, (void*)buf, len, 0);
-    if (-1 != bytes_sent)
+    int bytes_sent = send(platform->poll_fd.fd, (const char*)buf, (int)len, 0);
+    if (SOCKET_ERROR != bytes_sent)
     {
         rv = (size_t)bytes_sent;
         *errcode = 0;
@@ -76,11 +69,11 @@ size_t uxr_read_tcp_data_platform(struct uxrTCPPlatform* platform,
                                   uint8_t* errcode)
 {
     size_t rv = 0;
-    int poll_rv = poll(&platform->poll_fd, 1, timeout);
+    int poll_rv = WSAPoll(&platform->poll_fd, 1, timeout);
     if (0 < poll_rv)
     {
-        ssize_t bytes_received = recv(platform->poll_fd.fd, (void*)buf, len, 0);
-        if (-1 != bytes_received)
+        int bytes_received = recv(platform->poll_fd.fd, (char*)buf, (int)len, 0);
+        if (SOCKET_ERROR != bytes_received)
         {
             rv = (size_t)bytes_received;
             *errcode = 0;
@@ -99,6 +92,6 @@ size_t uxr_read_tcp_data_platform(struct uxrTCPPlatform* platform,
 
 void uxr_disconnect_tcp_platform(struct uxrTCPPlatform* platform)
 {
-    close(platform->poll_fd.fd);
-    platform->poll_fd.fd = -1;
+    closesocket(platform->poll_fd.fd);
+    platform->poll_fd.fd = INVALID_SOCKET;
 }
